+++ README.md
@@ -12,6 +12,16 @@
 
 ## 장바구니 기능 구현
 
+### 장바구니 기능
+
+- [ ] 장바구니에 상품 추가
+- [ ] 장바구니에 담긴 상품 제거
+- [ ] 장바구니 목록 조회
+
+### 장바구니 유저 인증
+
+- [ ] 사용자 정보는 요청 Header의 Authorization 필드를 사용해 인증 처리를 하여 얻는다.
+
 ## 장바구니 페이지 연동
 
 # 미션1 jwp-shopping-cart 기능목록
