+++ src/main/java/cart/auth/dao/UserDAO.java
@@ -1,19 +1,18 @@
-package cart.settings.dao;
+package cart.auth.dao;
 
-import cart.settings.domain.User;
-import cart.settings.dto.UserRequestDTO;
-
+import cart.auth.domain.User;
+import cart.auth.dto.UserRequestDTO;
 import java.util.List;
 
 public interface UserDAO {
-
+    
     boolean isExist(UserRequestDTO userRequestDTO);
-
+    
     User create(UserRequestDTO userRequestDTO);
-
+    
     User find(UserRequestDTO userRequestDTO);
-
+    
     List<User> findAll();
-
+    
     void delete(User user);
 }
