+++ src/main/java/cart/auth/domain/Password.java
@@ -1,4 +1,4 @@
-package cart.settings.domain;
+package cart.auth.domain;
 
 public class Password {
