+++ src/main/java/cart/catalog/controller/ExceptionAdvice.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
 public class ExceptionAdvice {
-
-    @ExceptionHandler({IllegalArgumentException.class})
-    public ResponseEntity<String> handle(final Exception exception) {
+    
+    @ExceptionHandler({Exception.class})
+    public ResponseEntity<String> handleAll(final Exception exception) {
         return ResponseEntity.badRequest().body(exception.getMessage());
     }
 }
