+++ src/main/java/cart/common/exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package cart.common.exceptions;
+
+public class NotFoundException extends RuntimeException {
+    
+    public NotFoundException(final String message) {
+        super(message);
+    }
+}
