+++ src/main/java/cart/settings/controller/SettingsController.java
@@ -1,22 +1,21 @@
 package cart.settings.controller;
 
-import cart.settings.dto.UserResponseDTO;
+import cart.auth.dto.UserResponseDTO;
 import cart.settings.service.SettingsService;
+import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import java.util.List;
-
 @Controller
 public class SettingsController {
-
+    
     private final SettingsService settingsService;
-
+    
     public SettingsController(final SettingsService settingsService) {
         this.settingsService = settingsService;
     }
-
+    
     @GetMapping("/settings")
     public String settings(final Model model) {
         final List<UserResponseDTO> allUsers = this.settingsService.findAllUsers();
