--- src/main/java/cart/settings/exceptions/NotFoundException.java
@@ -1,7 +1,0 @@
-package cart.settings.exceptions;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(final String item) {
-        super(item + " not found");
-    }
-}
